Add files via upload
This was my first attempt at learning the basic tags and syntax of HTML and CSS. I've chosen the general theme, of Travel in India, to implement all the topics that were taught in class. Even if the content may not make sense as a whole, this really helped me test the code hands on, which helped me understand the topics.